Function to predict using SuperSig object

Using a generated SuperSig, predict on a new dataset and return predicted probabilities for each observation.

predict_signature(object, newdata, factor)

Arguments

object
an object of class "SuperSig"

newdata
a data frame of mutations

factor
the factor/exposure (e.g. "age", "smoking")

Value

predict_signature returns the original data frame with additional columns for the feature counts and classification score

Examples

head(example_dt) # use example data from package

#>   sample_id age chromosome  position ref alt
#> 1         1  50       chr1  94447621   G   C
#> 2         1  50       chr2 202005395   A   C
#> 3         1  50       chr7  20784978   T   A
#> 4         1  50       chr7  87179255   C   G
#> 5         1  50      chr19   1059712   G   T
#> 6         2  55       chr1  76226977   T   C

input_dt <- make_matrix(example_dt) # convert to correct format
input_dt$IndVar <- c(1, 1, 1, 0, 0) # add IndVar column
head(input_dt)

#> # A tibble: 5 x 99
#>   sample_id   age `A[T>G]T` `C[T>A]A` `G[C>A]A` `G[C>G]G` `G[C>G]T` `A[C>G]T`
#>       <dbl> <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#> 1         1    50         1         1         1         1         1         0
#> 2         2    55         0         0         0         0         0         1
#> 3         3    72         1         1         0         0         0         0
#> 4         4    53         0         0         0         0         0         1
#> 5         5    48         0         1         1         0         0         0
#> # … with 91 more variables: `C[C>A]T` <dbl>, `C[T>C]G` <dbl>, `T[C>A]C` <dbl>,
#> #   `T[C>A]T` <dbl>, `A[T>C]C` <dbl>, `C[T>C]C` <dbl>, `T[T>A]C` <dbl>,
#> #   `C[C>G]G` <dbl>, `G[C>T]A` <dbl>, `A[C>A]T` <dbl>, `C[C>A]C` <dbl>,
#> #   `G[T>G]T` <dbl>, `C[C>T]C` <dbl>, `T[C>T]C` <dbl>, `A[C>T]C` <dbl>,
#> #   `G[C>T]C` <dbl>, `C[C>T]T` <dbl>, `T[C>T]T` <dbl>, `A[C>T]T` <dbl>,
#> #   `G[C>T]T` <dbl>, `C[C>T]A` <dbl>, `T[C>T]A` <dbl>, `A[C>T]A` <dbl>,
#> #   `C[C>T]G` <dbl>, `T[C>T]G` <dbl>, `A[C>T]G` <dbl>, `G[C>T]G` <dbl>,
#> #   `A[C>A]C` <dbl>, `G[C>A]C` <dbl>, `G[C>A]T` <dbl>, `C[C>A]A` <dbl>,
#> #   `T[C>A]A` <dbl>, `A[C>A]A` <dbl>, `C[C>A]G` <dbl>, `T[C>A]G` <dbl>,
#> #   `A[C>A]G` <dbl>, `G[C>A]G` <dbl>, `C[C>G]C` <dbl>, `T[C>G]C` <dbl>,
#> #   `A[C>G]C` <dbl>, `G[C>G]C` <dbl>, `C[C>G]T` <dbl>, `T[C>G]T` <dbl>,
#> #   `C[C>G]A` <dbl>, `T[C>G]A` <dbl>, `A[C>G]A` <dbl>, `G[C>G]A` <dbl>,
#> #   `T[C>G]G` <dbl>, `A[C>G]G` <dbl>, `T[T>C]C` <dbl>, `G[T>C]C` <dbl>,
#> #   `C[T>C]T` <dbl>, `T[T>C]T` <dbl>, `A[T>C]T` <dbl>, `G[T>C]T` <dbl>,
#> #   `C[T>C]A` <dbl>, `T[T>C]A` <dbl>, `A[T>C]A` <dbl>, `G[T>C]A` <dbl>,
#> #   `T[T>C]G` <dbl>, `A[T>C]G` <dbl>, `G[T>C]G` <dbl>, `C[T>A]C` <dbl>,
#> #   `A[T>A]C` <dbl>, `G[T>A]C` <dbl>, `C[T>A]T` <dbl>, `T[T>A]T` <dbl>,
#> #   `A[T>A]T` <dbl>, `G[T>A]T` <dbl>, `T[T>A]A` <dbl>, `A[T>A]A` <dbl>,
#> #   `G[T>A]A` <dbl>, `C[T>A]G` <dbl>, `T[T>A]G` <dbl>, `A[T>A]G` <dbl>,
#> #   `G[T>A]G` <dbl>, `C[T>G]C` <dbl>, `T[T>G]C` <dbl>, `A[T>G]C` <dbl>,
#> #   `G[T>G]C` <dbl>, `C[T>G]T` <dbl>, `T[T>G]T` <dbl>, `C[T>G]A` <dbl>,
#> #   `T[T>G]A` <dbl>, `A[T>G]A` <dbl>, `G[T>G]A` <dbl>, `C[T>G]G` <dbl>,
#> #   `T[T>G]G` <dbl>, `A[T>G]G` <dbl>, `G[T>G]G` <dbl>, IndVar <dbl>

out <- get_signature(data = input_dt, factor = "Age") # get SuperSig

#> Begin feature engineering...

#> Begin cross-validated selection over 3 features and 15 inner folds...

#> ...testing inner fold 1

#> ...testing inner fold 2

#> ...testing inner fold 3

#> ...testing inner fold 4

#> ...testing inner fold 5

#> ...testing inner fold 6

#> ...testing inner fold 7

#> ...testing inner fold 8

#> ...testing inner fold 9

#> ...testing inner fold 10

#> ...testing inner fold 11

#> ...testing inner fold 12

#> ...testing inner fold 13

#> ...testing inner fold 14

#> ...testing inner fold 15

#> $features_gmsa
#> $features_gmsa$new_partition
#> $features_gmsa$new_partition$X1
#>        F3 
#> "C[T>A]A" 
#> 
#> $features_gmsa$new_partition$X2
#>        F2 
#> "A[C>G]T" 
#> 
#> $features_gmsa$new_partition$X3
#>       F11       F12       F13       F14      F117      F118      F119      F132 
#> "A[C>A]A" "A[C>A]C" "A[C>A]G" "A[C>A]T" "A[C>G]A" "A[C>G]C" "A[C>G]G" "A[C>T]A" 
#>      F133      F134      F135      F148      F149      F150      F151      F163 
#> "A[C>T]C" "A[C>T]G" "A[C>T]T" "A[T>A]A" "A[T>A]C" "A[T>A]G" "A[T>A]T" "A[T>C]A" 
#>      F164      F165      F166      F179      F180      F181      F182       F15 
#> "A[T>C]C" "A[T>C]G" "A[T>C]T" "A[T>G]A" "A[T>G]C" "A[T>G]G" "A[T>G]T" "C[C>A]A" 
#>       F16       F17       F18      F120      F121      F122      F123      F136 
#> "C[C>A]C" "C[C>A]G" "C[C>A]T" "C[C>G]A" "C[C>G]C" "C[C>G]G" "C[C>G]T" "C[C>T]A" 
#>      F137      F138      F139      F152      F153      F154      F167      F168 
#> "C[C>T]C" "C[C>T]G" "C[C>T]T" "C[T>A]C" "C[T>A]G" "C[T>A]T" "C[T>C]A" "C[T>C]C" 
#>      F169      F170      F183      F184      F185      F186       F19      F110 
#> "C[T>C]G" "C[T>C]T" "C[T>G]A" "C[T>G]C" "C[T>G]G" "C[T>G]T" "G[C>A]A" "G[C>A]C" 
#>      F111      F112      F124      F125      F126      F127      F140      F141 
#> "G[C>A]G" "G[C>A]T" "G[C>G]A" "G[C>G]C" "G[C>G]G" "G[C>G]T" "G[C>T]A" "G[C>T]C" 
#>      F142      F143      F155      F156      F157      F158      F171      F172 
#> "G[C>T]G" "G[C>T]T" "G[T>A]A" "G[T>A]C" "G[T>A]G" "G[T>A]T" "G[T>C]A" "G[T>C]C" 
#>      F173      F174      F187      F188      F189      F190      F113      F114 
#> "G[T>C]G" "G[T>C]T" "G[T>G]A" "G[T>G]C" "G[T>G]G" "G[T>G]T" "T[C>A]A" "T[C>A]C" 
#>      F115      F116      F128      F129      F130      F131      F144      F145 
#> "T[C>A]G" "T[C>A]T" "T[C>G]A" "T[C>G]C" "T[C>G]G" "T[C>G]T" "T[C>T]A" "T[C>T]C" 
#>      F146      F147      F159      F160      F161      F162      F175      F176 
#> "T[C>T]G" "T[C>T]T" "T[T>A]A" "T[T>A]C" "T[T>A]G" "T[T>A]T" "T[T>C]A" "T[T>C]C" 
#>      F177      F178      F191      F192      F193      F194 
#> "T[T>C]G" "T[T>C]T" "T[T>G]A" "T[T>G]C" "T[T>G]G" "T[T>G]T" 
#> 
#> 
#> 
#> $features_selected
#> [1] "X1" "X2" "X3"
#> 
#> $new_partition
#> $new_partition$X1
#>        F3 
#> "C[T>A]A" 
#> 
#> $new_partition$X2
#>        F2 
#> "A[C>G]T" 
#> 
#> $new_partition$X3
#>       F11       F12       F13       F14      F117      F118      F119      F132 
#> "A[C>A]A" "A[C>A]C" "A[C>A]G" "A[C>A]T" "A[C>G]A" "A[C>G]C" "A[C>G]G" "A[C>T]A" 
#>      F133      F134      F135      F148      F149      F150      F151      F163 
#> "A[C>T]C" "A[C>T]G" "A[C>T]T" "A[T>A]A" "A[T>A]C" "A[T>A]G" "A[T>A]T" "A[T>C]A" 
#>      F164      F165      F166      F179      F180      F181      F182       F15 
#> "A[T>C]C" "A[T>C]G" "A[T>C]T" "A[T>G]A" "A[T>G]C" "A[T>G]G" "A[T>G]T" "C[C>A]A" 
#>       F16       F17       F18      F120      F121      F122      F123      F136 
#> "C[C>A]C" "C[C>A]G" "C[C>A]T" "C[C>G]A" "C[C>G]C" "C[C>G]G" "C[C>G]T" "C[C>T]A" 
#>      F137      F138      F139      F152      F153      F154      F167      F168 
#> "C[C>T]C" "C[C>T]G" "C[C>T]T" "C[T>A]C" "C[T>A]G" "C[T>A]T" "C[T>C]A" "C[T>C]C" 
#>      F169      F170      F183      F184      F185      F186       F19      F110 
#> "C[T>C]G" "C[T>C]T" "C[T>G]A" "C[T>G]C" "C[T>G]G" "C[T>G]T" "G[C>A]A" "G[C>A]C" 
#>      F111      F112      F124      F125      F126      F127      F140      F141 
#> "G[C>A]G" "G[C>A]T" "G[C>G]A" "G[C>G]C" "G[C>G]G" "G[C>G]T" "G[C>T]A" "G[C>T]C" 
#>      F142      F143      F155      F156      F157      F158      F171      F172 
#> "G[C>T]G" "G[C>T]T" "G[T>A]A" "G[T>A]C" "G[T>A]G" "G[T>A]T" "G[T>C]A" "G[T>C]C" 
#>      F173      F174      F187      F188      F189      F190      F113      F114 
#> "G[T>C]G" "G[T>C]T" "G[T>G]A" "G[T>G]C" "G[T>G]G" "G[T>G]T" "T[C>A]A" "T[C>A]C" 
#>      F115      F116      F128      F129      F130      F131      F144      F145 
#> "T[C>A]G" "T[C>A]T" "T[C>G]A" "T[C>G]C" "T[C>G]G" "T[C>G]T" "T[C>T]A" "T[C>T]C" 
#>      F146      F147      F159      F160      F161      F162      F175      F176 
#> "T[C>T]G" "T[C>T]T" "T[T>A]A" "T[T>A]C" "T[T>A]G" "T[T>A]T" "T[T>C]A" "T[T>C]C" 
#>      F177      F178      F191      F192      F193      F194 
#> "T[T>C]G" "T[T>C]T" "T[T>G]A" "T[T>G]C" "T[T>G]G" "T[T>G]T" 
#> 
#> 
#> $dt_new
#> # A tibble: 5 x 6
#>      X1    X2    X3 IndVar   AGE TOTAL_MUTATIONS
#>   <dbl> <dbl> <dbl>  <dbl> <dbl>           <dbl>
#> 1     1     0     4      1    50               5
#> 2     0     1     4      1    55               5
#> 3     1     0     4      1    72               5
#> 4     0     1     4      0    53               5
#> 5     1     0     4      0    48               5
#> 
#> $select_n
#> Logit 
#>     0 
#> 
#> $auc
#> Logit 
#>     0 
#> 
#> $signature
#> $signature$mean_diffs
#> # A tibble: 1 x 0
#> 
#> $signature$select_n
#> Logit 
#>     0 
#> 
#> 
#> $classifier
#> $classifier$Logit
#> 
#> Call:  glm(formula = IndVar ~ ., family = binomial(), data = x)
#> 
#> Coefficients:
#> (Intercept)  
#>      0.4055  
#> 
#> Degrees of Freedom: 4 Total (i.e. Null);  4 Residual
#> Null Deviance:	    6.73 
#> Residual Deviance: 6.73 	AIC: 8.73
#> 
#> 

newdata <- predict_signature(out, newdata = input_dt, factor = "age")
suppressPackageStartupMessages({library(dplyr)})
head(newdata)

#> # A tibble: 5 x 100
#>   sample_id   AGE `A[T>G]T` `C[T>A]A` `G[C>A]A` `G[C>G]G` `G[C>G]T` `A[C>G]T`
#>       <dbl> <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#> 1         1    50         1         1         1         1         1         0
#> 2         2    55         0         0         0         0         0         1
#> 3         3    72         1         1         0         0         0         0
#> 4         4    53         0         0         0         0         0         1
#> 5         5    48         0         1         1         0         0         0
#> # … with 92 more variables: `C[C>A]T` <dbl>, `C[T>C]G` <dbl>, `T[C>A]C` <dbl>,
#> #   `T[C>A]T` <dbl>, `A[T>C]C` <dbl>, `C[T>C]C` <dbl>, `T[T>A]C` <dbl>,
#> #   `C[C>G]G` <dbl>, `G[C>T]A` <dbl>, `A[C>A]T` <dbl>, `C[C>A]C` <dbl>,
#> #   `G[T>G]T` <dbl>, `C[C>T]C` <dbl>, `T[C>T]C` <dbl>, `A[C>T]C` <dbl>,
#> #   `G[C>T]C` <dbl>, `C[C>T]T` <dbl>, `T[C>T]T` <dbl>, `A[C>T]T` <dbl>,
#> #   `G[C>T]T` <dbl>, `C[C>T]A` <dbl>, `T[C>T]A` <dbl>, `A[C>T]A` <dbl>,
#> #   `C[C>T]G` <dbl>, `T[C>T]G` <dbl>, `A[C>T]G` <dbl>, `G[C>T]G` <dbl>,
#> #   `A[C>A]C` <dbl>, `G[C>A]C` <dbl>, `G[C>A]T` <dbl>, `C[C>A]A` <dbl>,
#> #   `T[C>A]A` <dbl>, `A[C>A]A` <dbl>, `C[C>A]G` <dbl>, `T[C>A]G` <dbl>,
#> #   `A[C>A]G` <dbl>, `G[C>A]G` <dbl>, `C[C>G]C` <dbl>, `T[C>G]C` <dbl>,
#> #   `A[C>G]C` <dbl>, `G[C>G]C` <dbl>, `C[C>G]T` <dbl>, `T[C>G]T` <dbl>,
#> #   `C[C>G]A` <dbl>, `T[C>G]A` <dbl>, `A[C>G]A` <dbl>, `G[C>G]A` <dbl>,
#> #   `T[C>G]G` <dbl>, `A[C>G]G` <dbl>, `T[T>C]C` <dbl>, `G[T>C]C` <dbl>,
#> #   `C[T>C]T` <dbl>, `T[T>C]T` <dbl>, `A[T>C]T` <dbl>, `G[T>C]T` <dbl>,
#> #   `C[T>C]A` <dbl>, `T[T>C]A` <dbl>, `A[T>C]A` <dbl>, `G[T>C]A` <dbl>,
#> #   `T[T>C]G` <dbl>, `A[T>C]G` <dbl>, `G[T>C]G` <dbl>, `C[T>A]C` <dbl>,
#> #   `A[T>A]C` <dbl>, `G[T>A]C` <dbl>, `C[T>A]T` <dbl>, `T[T>A]T` <dbl>,
#> #   `A[T>A]T` <dbl>, `G[T>A]T` <dbl>, `T[T>A]A` <dbl>, `A[T>A]A` <dbl>,
#> #   `G[T>A]A` <dbl>, `C[T>A]G` <dbl>, `T[T>A]G` <dbl>, `A[T>A]G` <dbl>,
#> #   `G[T>A]G` <dbl>, `C[T>G]C` <dbl>, `T[T>G]C` <dbl>, `A[T>G]C` <dbl>,
#> #   `G[T>G]C` <dbl>, `C[T>G]T` <dbl>, `T[T>G]T` <dbl>, `C[T>G]A` <dbl>,
#> #   `T[T>G]A` <dbl>, `A[T>G]A` <dbl>, `G[T>G]A` <dbl>, `C[T>G]G` <dbl>,
#> #   `T[T>G]G` <dbl>, `A[T>G]G` <dbl>, `G[T>G]G` <dbl>, IndVar <dbl>,
#> #   score <dbl>

head(newdata %>% select(X1, score))

#> Error: